% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{aperm,Grid-method}
\alias{aperm,Grid-method}
\title{Permutation method}
\usage{
\S4method{aperm}{Grid}(a, perm, ...)
}
\arguments{
\item{a}{Grid object}

\item{perm}{new order indexes}

\item{...}{Not used.}
}
\value{
Permuted Grid.
}
\description{
The generalised transpose changes the order of the factors 
   keeping the order of the observations. The transformation to a data frame is
   not efficient in term of computations, but is easily understood.
}
\author{
Yves Deville
}
