% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{as.data.frame,Grid-method}
\alias{as.data.frame,Grid-method}
\title{Coerce Grid in data.frame}
\usage{
\S4method{as.data.frame}{Grid}(x, jitter = FALSE, ...)
}
\arguments{
\item{x}{Grid object}

\item{jitter}{if TRUE, add a small white noise to data. Defaults to FALSE.}

\item{...}{Not used.}
}
\value{
Coerced data.frame object
}
\description{
Coerce Grid in data.frame
}
\author{
Yves Deville
}
