% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{as.Grid.list}
\alias{as.Grid.list}
\title{Coercion to \code{Grid}}
\usage{
\method{as.Grid}{list}(object, ...)
}
\arguments{
\item{object}{An object to be coerced into a \code{Grid} object.}

\item{...}{Not used yet.}
}
\value{
Object with S4 class \code{"Grid"}.
}
\description{
Coercion to \code{Grid}
}
