% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.Grid}
\alias{as.Grid}
\title{Coercion of objects to \code{Grid}}
\usage{
as.Grid(object, ...)
}
\arguments{
\item{object}{An object to be coerced to \code{Grid}, typically a
data frame or a matrix.}

\item{...}{Not used yet.}
}
\value{
An object with S4 class \code{"Grid"}.
}
\description{
Coercion of objects into \code{Grid}.
}

