% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{as.matrix,Grid-method}
\alias{as.matrix,Grid-method}
\title{Coerce Grid in matrix}
\usage{
\S4method{as.matrix}{Grid}(x, jitter = FALSE, ...)
}
\arguments{
\item{x}{Grid object}

\item{jitter}{if TRUE, add a small white noise to data. Defaults to FALSE.}

\item{...}{Not used.}
}
\value{
Coerced matrix object
}
\description{
Coerce Grid in matrix
}
\author{
Yves Deville
}
