% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{dim,Grid-method}
\alias{dim,Grid-method}
\title{Get dimension of given Grid}
\usage{
\S4method{dim}{Grid}(x)
}
\arguments{
\item{x}{Grid object}
}
\value{
Number of dimensions/columns of the Grid object.
}
\description{
Get dimension of given Grid
}
\author{
Yves Deville
}
