% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{dimnames,Grid-method}
\alias{dimnames,Grid-method}
\title{Get dimension names of given Grid}
\usage{
\S4method{dimnames}{Grid}(x)
}
\arguments{
\item{x}{Grid object}
}
\value{
Names of dimensions/columns of the Grid object.
}
\description{
Get dimension names of given Grid
}
\author{
Yves Deville
}
