% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{dimnames<-,Grid,ANY-method}
\alias{dimnames<-,Grid,ANY-method}
\alias{dimnames<-,Grid-method}
\title{Set dimension names of Grid}
\usage{
\S4method{dimnames}{Grid,ANY}(x) <- value
}
\arguments{
\item{x}{Grid object}

\item{value}{New dimension names of Grid}
}
\value{
Modified Grid object
}
\description{
Set dimension names of Grid
}
\author{
Yves Deville
}
