% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drop_Grid}
\alias{drop_Grid}
\title{Drop the dimensions with a unique level in a Grid
object}
\usage{
drop_Grid(x, ...)
}
\arguments{
\item{x}{An object with class \code{Grid}}

\item{...}{Not used yet.}
}
\value{
An object of class \code{Grid} but with the fixed
dimensions removed.
}
\description{
Drop the dimensions with a unique level in a \code{Grid} object.
}
\examples{
myGD <- Grid(nlevels = c(3, 1, 4, 1, 5))
myGD
plot(myGD)
myGD1 <- drop_Grid(myGD)
}
\seealso{
The usual \code{\link{drop}} function for arrays.
}

