% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\docType{class}
\name{Grid-class}
\alias{Grid-class}
\title{S4 class for Grid}
\description{
This class is intended handle Grid object,
    with support for useful methods like levels, ...
}
\section{Slots}{

\describe{
\item{\code{dim}}{number of dimensions}

\item{\code{dimNames}}{name of dimensions}

\item{\code{levels}}{cardinal of discrete values}

\item{\code{index}}{}
}}

\author{
Yves Deville
}
