% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{length,Grid-method}
\alias{length,Grid-method}
\title{Get length of Grid}
\usage{
\S4method{length}{Grid}(x)
}
\arguments{
\item{x}{Grid object}
}
\value{
Length of the Grid object.
}
\description{
Get length of Grid
}
\author{
Yves Deville
}
