% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{levels,Grid-method}
\alias{levels,Grid-method}
\title{Get levels of Grid}
\usage{
\S4method{levels}{Grid}(x)
}
\arguments{
\item{x}{Grid object}
}
\value{
Levels of the Grid object.
}
\description{
Get levels of Grid
}
\author{
Yves Deville
}
