% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{nlevels,Grid-method}
\alias{nlevels,Grid-method}
\title{Get number of levels of Grid}
\usage{
\S4method{nlevels}{Grid}(x)
}
\arguments{
\item{x}{Grid object}
}
\value{
Number of levels of the Grid object.
}
\description{
Get number of levels of Grid
}
\author{
Yves Deville
}
