% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{plot,Grid-method}
\alias{plot,Grid-method}
\title{Plot Grid as pairs}
\usage{
\S4method{plot}{Grid}(x, y, jitter = FALSE, ...)
}
\arguments{
\item{x}{Grid object}

\item{y}{optional data to add in plot alongside x (should have same length)}

\item{jitter}{if TRUE, add a small white noise to data. Defaults to FALSE.}

\item{...}{Optional arguments passed to 'pairs()'}
}
\value{
'pairs' object
}
\description{
Plot Grid as pairs
}
\author{
Yves Deville
}
