% Generated by roxygen2 (4.0.2): do not edit by hand
\name{range_Grid}
\alias{range_Grid}
\title{Range of a \code{Grid} object}
\usage{
range_Grid(X)
}
\arguments{
\item{X}{An object with S4 class \code{"Grid"}, or an object
with S3 class \code{"matrix"} or \code{"data.frame"} that can
be coerced into a  \code{"Grid"} object.}
}
\value{
A numeric matrix with two rows representing the min
and max of the levels for each dimension of the grid.
}
\description{
Range of a \code{Grid} object.
}
\examples{
set.seed(1234)
myGD <- randGrid(dim = 4)
range_Grid(myGD)
range_Grid(as.data.frame(myGD))
range_Grid(as.matrix(myGD))
}

