% Generated by roxygen2 (4.0.2): do not edit by hand
\name{round_levels}
\alias{round_levels}
\title{Round a vector (of levels)}
\usage{
round_levels(x)
}
\arguments{
\item{x}{Numeric vector to be rounded.}
}
\value{
Rounded numeric vector.
}
\description{
Round a vector (of levels).
}
\note{
At the time, this function can produce poor results
when the relative variation in \code{x} is small.
}
\examples{
## good
round_levels((1:3) / 10^5)
round_levels( 10 + (1:3) / 10^5)
## bad
round_levels( 10 + (1:3) / 10^6)
}

