% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{sampleIn,data.frame-method}
\alias{sampleIn,data.frame-method}
\title{Sample in data.frame}
\usage{
\S4method{sampleIn}{data.frame}(X, size = 1L, atSample = FALSE, ...)
}
\arguments{
\item{X}{design data.frame}

\item{size}{Number of sample points.}

\item{atSample}{should new sample points obtained using \code{sample}}

\item{...}{optional args passed to 'sample()'}
}
\description{
Sample in data.frame
}
