% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{sampleIn,Grid-method}
\alias{sampleIn,Grid-method}
\title{Sample in Grid}
\usage{
\S4method{sampleIn}{Grid}(X, size = 1L, atSample = FALSE, ...)
}
\arguments{
\item{X}{design Grid}

\item{size}{Number of sample points.}

\item{atSample}{should new sample points obtained using \code{sample}}

\item{...}{optional args passed to 'sample()'}
}
\description{
Sample in Grid
}
