% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{sampleIn,matrix-method}
\alias{sampleIn,matrix-method}
\title{Sample in matrix}
\usage{
\S4method{sampleIn}{matrix}(X, size = 1L, atSample = FALSE, ...)
}
\arguments{
\item{X}{design matrix}

\item{size}{Number of sample points.}

\item{atSample}{should new sample points obtained using \code{sample}}

\item{...}{optional args passed to 'sample()'}
}
\description{
Sample in matrix
}
