% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{sampleIn}
\alias{sampleIn}
\alias{sampleIn,matrix-method}
\alias{sampleIn,data.frame-method}
\alias{sampleIn,Grid-method}
\title{Sample from/in a design object}
\usage{
sampleIn(X, size = 1L, atSample = FALSE, ...)

\S4method{sampleIn}{matrix}(X, size = 1L, atSample = FALSE, ...)

\S4method{sampleIn}{data.frame}(X, size = 1L, atSample = FALSE, ...)

\S4method{sampleIn}{Grid}(X, size = 1L, atSample = FALSE, ...)
}
\arguments{
\item{X}{design Grid}

\item{size}{Number of sample points.}

\item{atSample}{should new sample points obtained using \code{sample}}

\item{...}{optional args passed to 'sample()'}
}
\value{
A matrix with \code{size} rows. When \code{atSample} is
\code{TRUE}, the matrix has an attribute named \code{index} which
gives the position of the sampled items in the original grid.
}
\description{
Sample from/in a design object.
}
\details{
From the object, \code{size} 'new' design points are sampled or
drawn. When \code{atSample} is \code{FALSE}, the range of each
column is computed or extracted from the object \code{x}, and then
independent drawings from the corresponding uniform distributions
are used.
}
\note{
When \code{x} is of class \code{"Grid"}, it may be the case
that the drawings with \code{atSample = FALSE} have no meaning since some
of the variables may be discrete variables and not continuous ones.
}
\examples{
g <- Grid(levels = list("Temp" = c(400, 450, 500),
                        "Bore" = c(0, 600, 1200, 1800, 2400)))
X1 <- sampleIn(g, size = 4)
X2 <- sampleIn(g, size = 4, atSample = TRUE)

## this must be zero
sum(abs(as.matrix(g)[attr(X2, "index"), ] - X2))



}
\seealso{
The \code{\link{sample}} function.
}
