% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridData.R
\name{show,Grid-method}
\alias{show,Grid-method}
\title{Print Grid object informations}
\usage{
\S4method{show}{Grid}(object)
}
\arguments{
\item{object}{Grid object}
}
\value{
Nothing.
}
\description{
Print Grid object informations
}
\author{
Yves Deville
}
