\docType{package}
\name{smint-package}
\alias{smint-package}
\alias{smint}
\encoding{UTF-8}

\title{

  Multivariate Interpolation for Scattered Data and Gridded data

}
\description{

  This package implements several methods for Multivariate Interpolation
  for Scattered Data and Gridded data with large number of interpolation
  nodes. The target dimension runs from 4 to 8.
  
}
\details{
  
  \tabular{ll}{
    Package: \tab smint\cr
    Type: \tab Package\cr
    Version: \tab 0.4-2\cr
    Date: \tab 2016-01-14\cr
    License: \tab GPL-3\cr
    LazyLoad: \tab yes\cr
  }

}
\author{
  
  Yves Deville <deville.yves(at)alpestat.com> and Yann Richet
  <yann.richet(at)irsn.fr>.

  This package embeds a Fortran code written by Alain Hébert. It also
  embeds codes from the SHEPPACK Fortran package written by William
  I. Thacker, Jingwei Zhang, Layne T. Watson, Jeffrey B. Birch, Manjula
  A. Iyer, Michael W. Berry, and adapted from previous code by Robert
  Lenka.

}
\references{

  W. Cheney and W. Light W. (2009) \emph{A course in approximation
  theory}, American Mathematical Society (AMS).

  C. de Boor (2001) \emph{A Practical Guide to Splines}, revis. Ed.,
  Springer-Verlag.
  
  W.I. Thacker, J. Zhang, L.T. Watson, J.B. Birch, M.A. Iyer and
  M.W. Berry (2010). Algorithm 905: SHEPPACK: Modified Shepard Algorithm
  for Interpolation of Scattered Multivariate Data \emph{ACM Trans. on
  Math. Software} (TOMS) Vol. 37, n. 3.
  \href{http://dl.acm.org/citation.cfm?id=1824812}{link}

}
%% \examples{}